Align status of Entries API docs with implementation reality
For all articles for features of the File and Directory Entries API,
this change drops all flags and statements indicating the API is
deprecated, non-standard, or experimental.

The API has now been implemented and shipped for a long time now in all
current browser engines — so at this point it’s not helpful to
developers to warn them about it being unstable or non-standard.

This change also makes a variety of other changes to align with the
current specification — including:

- replacing references to the obsolete FileError object with references
   to the DOMException object instead

- dropping any mentions of the former parts of the specification that
  had to do with writing (creating) files (the current spec only defines
  an API for reading files)

- dropping any mentions of the former parts of the specification that
  defines synchronous equivalents for some of the API members

And the change also includes some fixes to un-conflate the File and
Directory Entries API from the previous, ancient, non-standard File
System API.

Fixes https://github.com/mdn/content/issues/8263