Fix incorrect response in HTTP Auth diagram
The original sequence diagram had an incorrect HTTP response type of '401 Forbidden'. It should be either '403 Forbidden' or '401 Unauthorized'. From the rfc2617 spec (https://datatracker.ietf.org/doc/html/rfc2617):
"If the origin server does not wish to accept the credentials sent with a request, it SHOULD return a 401 (Unauthorized) response."

The new image updates the response type for invalid credentials to be '401 Unauthorized' and also makes the text a bit clearer to read