Add .gitpod.yml
This should help Gitpod set things up as described in README.md
(assuming Gitpod is opened on the user's own fork).

Unfortunately, there doesn't seem to be a way to refer to a user's fork
unless you already know where it is, so it doesn't look like we can
add a link to open Gitpod in this way to README.md.