Make the note less ambiguous
The specification actually doesn't _require_ `border-style` and/or `border-width` for `border-image` to be rendered (https://github.com/w3c/csswg-drafts/issues/655#issuecomment-365454403). However, some browsers indeed would not render it without them (although this is technically a bug and is being fixed), so specifying them is a reasonable practical requirement at the moment (and providing fallbacks is a good practice anyway).