Added very important yet missing information. (#8691)
The writing direction sensitivity of the direction of rotation seems to be
an unspecified feature whose practical implications are so obviously
appropriate that it seems to have been consistently yet silently
implemented. Tested in Chrome and Firefox. However, its effect is sometimes
unexpected, like when angle quotation marks are repurposed as twisties (and
their bidi-mirrored=yes property does not fix it).