element.getBoundingClientRect: Target purple example (#9100)
I would expect the 2nd live example to show the bounding rectangle of the purple example div (just like the 1st live example) instead of the empty `div` before it. Unsure where this empty `div` is coming from but without it, the JavaScript would actually target the correct one. Wondering if it was added by mistake at some point.