Change Checkbox_radio_button example
Changes for `Checkbox_radio_button` example to conform to the rule stated at the top of the page (i.e. _<input type="radio"> elements, when all radio buttons with the same name value in the form are unchecked_) that is we need more than one radio button in order to see the effect. Moreover, minor changes have been made.