Update index.md
Not sure how to phrase it, but the page should say that execCommand is only partially deprecated. The command `insertText` was lately implemented in FF89 and is available in all other browsers (except IE).

See: https://github.com/w3c/editing/issues/160#issuecomment-806165950

Should probably also mention this in Compat.