RegExp Assertions: fix note about ^ in group
There is a note in the Meaning column for character `^` which says:

> This character has a different meaning when it appears at the start of a group.

I am unable to find any difference in meaning when ^ is used in a group,
at the start or otherwise.  For example, `^hi` and `(^hi)` match exactly
the same strings as far as I am aware.

I suspect this note is referring to the different behavior when used in
[CharacterClass], referred to as ranges elsewhere in the docs, and in
the "Different meaning of '?!' combination usage in Assertions and
Ranges" section of the same file.

Fix the note to refer to "a range" instead of "a group".

[CharacterClass]: https://tc39.es/ecma262/multipage/text-processing.html#sec-characterclass

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>