Update Firefox Dev Tools - Minor Text Change
A minor change. When the article covers the main tools (that have the screenshots such as Web Console and JavaScript Debugger) I noticed that the description for that tool is after the screenshot. The usability issues with this are:
(1) When a newish user see's the screenshot they won't really know what they are looking at until they look at the text after the screenshot.
(2) It is unclear, at first, if the descriptive text belongs to the screenshot below or after the text.

I just copied and pasted the descriptive text (in the <p> tag) from the bottom of the image to the top of the image right under the title. You should find that more users will understand immediately what they are look at when they can read the descriptive text right before starring at the image. If anything the descriptive text can remind some users what the panel is for so they remember what they are looking at.