Update MutationObserver example
Currently the example found at https://codepen.io/webgeeker/full/YjrZgg/ doesn't work; instead of resizing, it maintains the same size and will show a scroll bar when the size of the contenteditable list exceeds the initial size. 

I ported the code over from [the blog post that I believe this references](https://hacks.mozilla.org/2012/05/dom-mutationobserver-reacting-to-dom-changes-without-killing-browser-performance/) and is referenced to in the section directly below so that the live example would be working as expected.