Clarify that Object is below null in the prototype chain
When I first read "Nearly all objects in JavaScript are instances of Object **which sits on the top of a prototype chain**" (emphasis added), I was confused, because the previous sentence states that `null` is the final link in prototype chains. I had to open up the browser console and look at {}.__proto__.__proto__ to see that `null` is indeed the prototype of Object, rather than this being an exception to `null` being the final link.

This edit clarifies that point—that most JavaScript objects have Object at the top of their prototype chain, but still just below null.