Make mouse following animation easier to read.
This example, as of now, is full of single-character variable names and a few unused functions. I understand that this code was, according to the original github repo, written as a POF and, understandably, not optimized. However, for learning purposes, making it easier to read in general would be more beneficial for everyone, especially beginners or those who are new to canvas. 

What I did:
1. Rewrote variable and function names like c, cc, a, gc, cn, etc. to be more self-explanatory.
2. Removed unused lines of code.
3. Moved script tag from top to bottom. 
4. window.onresize was removed; it was just generating random positions on every resize and causes flashing. 
5. window.onload removed and functions were moved outside.

All functionalities are the same after refactoring: https://codesandbox.io/s/mdn-mouse-following-edit-l5qd6?file=/index.html