Add information about drag feedback image
This is a contribution to the documentation for situations when an HTML element other than an image is used for a drag feedback image. It is important to note that an existing HTML element will not be usable if it is hidden outside the viewport via "display: none", absolute position or "visibility: hidden".