Issue with "Client-side storage"
Using version '1' to open the 'indexedDB' [1] kept throwing the
error of not knowing the object store name [2]. However, it
worked just fine in chromium. Maybe it was because I have already
tried executing the `js` code as I was going with the tutorial,
and did not use the complete `js` file from the
beginning. Anyways, I came up with this solution: just upgrade
the database version.

Footnotes:
----------
[1] let request = window.indexedDB.open('notes_db', 1);
[2] Uncaught DOMException: IDBDatabase.transaction: ‘notes_os’ is not
    a known object store name