Replace references to "void"
For a web developer audience, the term "void" does not describe a
value--it is the name of a rarely-used unary operator which happens to
produce the value `undefined`. This makes its use in the context of
function return values somewhat confusing. Replace the term "void" with
the term "undefined", and use the project's cross-linking capabilities
to relate this to the canonical definition of the "undefined" language
value.