array.concat() example improvement
The first example on the page for `Array.concat` was incorrectly understood by a colleague, believing the "results" in the comment describe the new value of `letters`. This small change should make it both clearer that the result is a return value, as well as make it more consistent with the other examples (where the results of `concat()` are also stored in an array and then logged).