Reword this comment
I just find the hyphenation a bit confusing: `page-script-added` sounds like a property, maybe, something that exists, but can't be seen. But that complete token never appears in the page or the page's JS. The first hyphen is a word separator, there is an implied `[.js]`, and the second hyphen is so that we can tack on a suffix, but the change in the hyphen just boggles my mind. I think this is a slightly clearer wording that shows we're clearly referring to the above page, and not some specific property or variable it might have added.