2x is missing
"2x" is missing in the srcset value at the end, otherwise it won't work. Also the live example on 
https://yari-demos.prod.mdn.mozit.cloud/en-US/docs/Web/API/HTMLImageElement/srcset/_sample_.Example.html
doesn't work because of the missing "2x" in the code. 
I tried in XCode Simulator with IPhone and Firefox Inspector in Responsiv Design Mode (DRP:2). As soon as I add the "2x" it works.