Line 61 - 63
on line 61 it suggests to see a demo by following a link.  The link is a zip file.  May I suggest it may be more appropriate to link to a codepen instead of downloading a zip file that downloads unknown code to my personal computer?  I know I would feel more comfortable with something like that.