Fix statement about allowed Content-Type values
The page claims the following:

> The only allowed values for the `Content-Type` header are:
> * `application/x-www-form-urlencoded`
> * `multipart/form-data`
> * `text/plain`

However, this statement is imprecise; for instance, a value of `text/plain; application/json` is allowed and does _not_ (all other things being equal) trigger a preflight request. For more about this, refer to [the relevant section of the Fetch standard](https://fetch.spec.whatwg.org/#ref-for-mime-type-essence).

This statement may mislead practitioners and lull them into a false sense of security in the face of cross-origin attacks. In particular, practitioners may believe that finding a match for `application/json` in the value of the `Content-Type` request header is a sufficient indicator that the request was preflighted.