Correct page about Access-Control-Expose-Headers
[The page about the `Access-Control-Expose-Headers` response header][aceh] has this to say about the wildcard directive:

> The value "`*`" only counts as a special wildcard value for requests without credentials (requests without HTTP cookies or HTTP authentication information). In requests with credentials, it is treated as the literal header name "`*`" without special semantics. **Note that the `Authorization` header can't be wildcarded and always needs to be listed explicitly.**

(my emphasis)

It's true that the [Fetch standard][fetch-std] lists `Authorization` as the sole _CORS non-wildcard request-header name_; accordinly, `Authorization` must indeed be explicitly allowed in the  `Access-Control-Allow-Headers` response header.

However, `Authorization` is a _request_ header, not a _response_ header, and the Fetch standard doesn't mention any special treatment of a header named `Authorization` in the context of the `Access-Control-Expose-Headers` response header. Therefore, the sentence I've emphasised above is out of place. It was most likely added as a result of a copy/paste error from [the page about `Access-Control-Allow-Headers` response header][acah], and it should be removed.

[acah]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Headers#directives
[aceh]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Expose-Headers#directives
[fetch-std]: https://fetch.spec.whatwg.org/#cors-non-wildcard-request-header-name