Update in Epilepsy area
Some recent WCAG discussion on the subject has led to revisiting this information in a modern light — The EFA recommendation from 2005 pre-dates smart phones, and was based on television set usage, defining a flash that "occupies a solid visual angle of ≥0.006 steradians (∼10% of the central visual field or 25% of screen area at typical viewing distances)."

This is essentially equivalent to a visual angle of 5°, which is easier to understand than the more arcane steradians. But there's more:

Since then, mobile devices have substantially changed how we view media. For a typical smart phone viewed very close (5"-6" or 12cm-15cm) that would be an area about 50px to 70px square.

This important change is being incorporated here, as the author of this document (myself) would like to ensure best, actionable guidance on this important topic.

Thank you

Andrew Somers