Change "validate" to "check"
## Changes made

It appears that the script `setupTypeScript.js` has been changed so the script in `package.json` formally named 'validate' is now named 'check'. Personally, I like the word validate better but am proposing changes to the tutorial that reflect the script as it currently is.

Code from current script. Starts on line 33.
```js
// Add script for checking
packageJSON.scripts = Object.assign(packageJSON.scripts, {
  "check": "svelte-check --tsconfig ./tsconfig.json"
})
```



## Other Possible Changes

There are a couple of place where I was unsure if the word "validate" should be changed to "check".

**bold emphasis added just temporary to highlight the difference.**

Don't know if you would want to change the word "validate" to "check" in this context as `check` is doing validation.

(1)
```
"Now we will undo these changes in order to start porting our application to TypeScript, so we won't be bothered with all the **validate** warnings." 
```
_(or maybe say "validation warnings" to avoid confusion)_

It could be changed to:

```
"Now we will undo these changes in order to start porting our application to TypeScript, so we won't be bothered with all the **check** warnings."
```

---
(2)

`validate` seems to mean something different here but wanted to double check that.

"Now if we try to create a `localStore` with something that cannot be converted to JSON via `JSON.stringify()`, for example an object with a function as a property, VS Code/`validate` will complain about it:"

--

I'm happy to make the above changes if desired.