MediaQueryList removeListener depracation
Affected page: https://developer.mozilla.org/en-US/docs/Web/CSS/Media_Queries/Testing_media_queries

The previous example is using deprecated `addListener` in text and `removeListener`  in both text and example code. Updated to use `addEventListener` and `removeEventListener` consistently. 

Notice for older browsers and older implementation is referenced in the text already but the example was not consistent with the text.