Update index.md where link is broken
The external link leads to a page or article that no longer exists. I don't know what the page/article was about, but the link text about "advice on handling email clients" doesn't make any sense to me and doesn't fit to the MDN page.
That's why I just removed the link because otherwise I couldn't find the article on the page either.