Fixed Incorrect Calculations (#11735)
`flex-grow` lets a flex item specify how it's main-size can grow when their is free space available in the parent container. When `flex-grow` is a positive number, it specifies the proportion of the free space that can be added to the item's main size. The existing calculations were made on the total space instead of free space. Here is a [Codepen](https://codepen.io/vthukral94/pen/NWazPmN) to justify new numbers.