Remove misleading `attr()` from the Syntax section (#10806)
I couldn't find anywhere in the documentation that `attr()` is allowed in the `clamp()`. 
I believe it'd be better off without this misleading mention of the `attr()` then. 
Would be happy to be proved wrong. 
I have also tried myself to use `attr()` inside the `clamp()` and it didn't work.