Changed the examples in the non-example sections
Changed the examples outside of the Example because the link was wider than the web page so the end was cut off as it didn't wrap. 
Changed from border to outline because dynamically changing a box model property is not a good example.
Second example was changed to background-color because that is likely something someone would want to do.