Added intro description of what a Proxy is
It started out with how to use it, without much of a clue as to what it does or why you would want to use it.  These objects should always start with a simplie description, so the reader doesn't have to read further to see if they care about it, or if it's what they're looking for.
Also converted some passive voice to active voice.