note that promise.all preserves order
I've had to google this a few times and I would like to add it to the MDN docs.

When passing an array of promises, `.all` will resolve them in order. This is guaranteed according to things I've found around the when searching, and having this in MDN docs will be helpful for the broader community.
https://stackoverflow.com/a/28066851
https://www.karltarvas.com/2015/01/21/order-of-resolved-values-in-promise-all.html

This is my first time contributing to MDN and thus am very happy to make any changes to the proposed change. I just want this information added in some capacity.