Add definition for the second custom element
The example only included definition for the first custom element
`<open-shadow>` while assuming that the reader knows what the
second `<closed-shadow>` elements definition looks like.

The two definitions are very similar but it is better to include
the full code that is required to run the example so the reader can
easily copy-paste it.