Fix the polyfill of Math.fround.
As shown here:
https://gist.github.com/shicks/7a97ec6b3f10212e60a89a7f6d2d097d?permalink_comment_id=4048135#gistcomment-4048135
the old polyfill was incorrect, because it was breaking ties *upwards*
when rounding. The spec of `fround` requires to break ties to *even*
instead.

The new polyfill correctly breaks ties to even, even at the critical
points of transitioning from subnormal to normal forms, and from
finite to infinity.

To perform ties-to-even, the code relies on the natural rounding
that will happen when getting into the subnormal forms of
(64-bit) `number`s.

In addition, the new polyfill does not assume that `Math.log` will
always return the best approximation. The spec of `Math.log`
explicitly says that it may not return the best approximation.