Update index.md with some changes for consistency (#12551)
* Update index.md with some changes for consistency

Compared all assessment pages (from the HTML and CSS modules so far) and found some small differences (content-wise) in the "Assessment or further help" sections. To keep everything consistent, I used the variant that currently appears on most of this pages (so that it also doesn't need so many changes) and adjusted it on those pages where there were small differences (different wording for example).
Additionally I added a hints and tips section, as it is (almost) always the case and I personally find it useful. Hope this fits, otherwise please let me know!