Update index.md where contents were swapped
If you follow the assessment page of most of the other HTML and CSS modules, the order of the "starting point" is as follows: text according to the assessment, alternative code editors and then the note on further help. That's why I have changed it here also in this way.