Un-break the live sample in “Taking still photos with WebRTC”
Fixes https://github.com/mdn/content/issues/11212

This is a follow-up to https://github.com/mdn/content/pull/12633. We
need to ensure we get the user to first open the (framed) live sample
in its own tab/window. Otherwise, it won’t be able to request permission
to access the camera.