update presentation role document
This update brings the none and presentation role intros more inline.  Calling out the specific examples of headings and lists in the intro to this page was not particularly clear, and is better explained in the body of the page's content.

I also changed the example from an `<img src=... role=none alt="">` to an `<hr role=none>`.  The image example was not really a good use of the role, because `alt=""` effectively already set the element to presentational, so it was redundant to specify the role as well.