small correction
In javascript | Mdn -> Javascript First Steps -> What went wrong? article, In the section "SyntaxError: missing ; before statement" the userGuess variable is defined as const according to "number-game-error.html", but, in the section it is defined as let and there is no relation of variable declaration in this section. The section talks only about error due to equality sign.