Change "creates a property 'a' on p"
The property is already created because the object inherits from "o". If after the inheritance we put console.log(p.a) it will give us the value of 2. I was reading it and I was confused, because I thought it haven't taken that value (that was kinda weird) so I had to check in the console and it copies the properties of all the object. I think it would be better to say assign because the property already exists. Hope it can help.