Insert reference-style links
This commit applies a pattern to reduce repetition and hopefully make maintenance easier.

For links to sections in the page, a reference has been added under the relevant header and all instances of the link replaced with the reference. If headers change in future, only the reference needs to be updated rather than every individual link. Having the reference adjacent to the header also acts as a prompt to make the necessary change.

References have only been created for headers which were identified as linked to within this page, when {{anch}} calls were replaced with MD links.

This commit is only a suggested approach and can be rolled back if the preference is to keep standard links throughout.