small change in topic "Accessing every item"
In topic, "Accessing every item" , we are printing the string whose length is greater than 8 characters long. But, In theory they mentioned that " The code below takes an array of strings and filters out all strings that are less than 8 characters long:". Instead of "less" word "greater" word is correct.

Link : https://developer.mozilla.org/en-US/docs/Learn/JavaScript/First_steps/Arrays