Remove 'Test skills' from 1st article abt strings
The 'Handling text - strings in JavaScript' has a link to 'Test Your Skills' for Strings. 

However all the tests require knowledge of basic string methods which is in the *next* article. So keeping 'Test Your Skills' before that confuses the student. Hence suggesting that it be removed from this 1st article about strings.