Add Examples for using the function constructor
The docs do not contain examples for how to use the function constructor to create new functions when passing in a function body. This commit fixes this.

Examples have been added to show how to pass in both function declarations and expressions to new Function and how to call the function created.