Improve the fallback example for :focus-visible (#13639)
In the overriding selector, the `:not()` pseudo-class increases specificity, which can cause problems with other styles on `:hover`, `:active`, and other pseudo-classes.
And in the case of `@supports`, the specificity remains the same as for simple pseudo-class selectors.
In addition, the `@supports` option is easier to read. And [it works exactly the same](https://codepen.io/firefoxic/pen/LYOoGJY).