Update accessability section
The "accessability concerns" section references an issue which was fixed in Chrome 90, as if it applied to current browser versions. The reference given for the issue itself - https://marcysutton.com/content-visibility-accessible-semantics - has been updated to note that it's no longer an issue. 

This describes the current behavior, while noting the past behavior prior to Chrome 90. Follow's Google's own description - https://web.dev/content-visibility/#a-note-on-accessibility