Adjust example directory structure authentication
The registration directory should be inside the templates directory, but that's not reflected in the example. Visually it looked like registration should be placed in the Django project folder directly (which is not the case).
Since I think the original quote should be kept, I propose to add a few non-breaking space HTML entities to signify the correct place of the registration directory.