Edit grammar
The first change makes it clear immediately that readers should direct their attention to the next example. Previously, if you were reading quickly it was easy to skim over this direction because it was at the end of the sentence, creating a few seconds of confusion.

The second change fixes an awkward construction.