Removed hidden from "passwordrules" div
When copying the html into an online sandbox, the tooltip would not pop up when hidden was is in the "passwordrules" div. Taking it out allowed the tooltip to appear on hover and disappear when not hovering. Not sure if it's a browser-specific issue, I'm using Chrome Version 99.0.4844.83 (Official Build) (x86_64) on macOS Version 12.2.1.