Doc: Fix confusing endIndex description.
'index _before_ which to end extraction' was confusing. Since endIndex is not the index _before_ the last char you want to extract, but the index _after_ the last char you want to extract. The user is originally oriented towards the last char they want, not the one after that. The new description is now also shorter and equal to the 'endIndex' description in '.slice()', so it's easier to compare the two and realize that they refer to the same index.