Doc: Fix confusing end index description for Array.prototype.slice()
'index _before_ which to end extraction' was a confusing indirection. Since end index is not the index _before_ the last char you want to extract, but the index _after_ the last char you want to extract. The new description is now equal to the 'endIndex' descriptions in 'String.prototype.substring()' as well as the latest doc for `String.prototype.slice()` (commit 3b21c37 in PR #14646), so it's easier to carry knowledge across.