Make GeneratorFunction doc be clear it’s not a property of global object
Fixes https://github.com/mdn/content/issues/14781

GeneratorFunction is an intrinsic object in the ES spec but it’s not a
exposed as a property of the global object. So there is no exposed
constructor with the name GeneratorFunction(), and we should make that
clear in the article. And because there’s no such exposed constructor,
we should not include a Syntax section for it.