Update Intl.ListFormat() params section + add more examples
Fixes https://github.com/mdn/content/issues/15167

This drops the statement “When `style` is "`short`" or "`narrow`",
"`unit`" is the only allowed value for the type option” — which doesn’t
align with the spec nor with implementations (that wording/restriction
was in the original proposal but ultimately didn’t make its way into the
spec. This also adds a lot more examples.