Remove outdated information about `xlink:href`
This warning suggests that `xlink:href` may still be necessary for cross-browser compatibility; however, the table at the bottom shows that the stable version of all browsers (including IE) now support `href`, so this warning is now misleading. In addition, it conflicts with the more accurate warning further down the page that `xlink:href` is deprecated.