Change CSS for nested `kbd` sample.
Currently, a class is used to target the nested `<kbd>` elements. I think this defeats the purpose of nesting.

Instead, I have removed the class and changed the selector to `kbd>kbd` which is more in keeping with nesting.

Can someone tell me what to do about updating the sample result?