Add a footnote to explain potential confusion (#14064)
* Add a footnote to explain potential confusion 

In the following tutorial user is tasked to add 3 to 74 (which type is string). First we add 3 to string so the result is 743. Tutorial explains that we need to change data type in the operation to get desirable result. After changing type in calculation output is 746, not 77. This may confuse the user to think that he made something wrong, in tutorial we wanted to get 77.

* Update files/en-us/learn/javascript/first_steps/math/index.md

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Update files/en-us/learn/javascript/first_steps/math/index.md

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

Co-authored-by: wbamberg <will@bootbonnet.ca>
Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>