A note for empty string returned for default ports (#15165)
* A note for empty string returned for default ports

Currently it's only noted that an empty string is returned when no port is included in the given url. This does not cover when a port is given, but with a protocol that has a matching default port. So `new URL('http://localhost:80').port` returns `''`, same with `new URL('https://localhost:443').port`, `new URL('ssh://localhost:22').port`, etc. Not sure if an example showing this should be added in the Examples section as well, but my assumption is that the note in the description is enough.

Also happy to take feedback on the wording, not sure if `given port number` is too explicit or `protocol's default` is too ambiguous.

* updated the URL.port description

- updated the description of how the port is parsed based on PR review
- only change was to change how the empty string is references based on other pages

* Copy edits

Co-authored-by: Michael[tm] Smith <mike@w3.org>