Remove comments about complex selector
I think it's time we could remove those comments, at least from the Example. According to [caniuse](https://caniuse.com/?search=not) it is pretty well supported.

As regarding the point in the description (i.e. _Using two selectors at the same time might not be supported in all browsers. Example: `:not(.foo, .bar)`. For wider support you could use, `:not(.foo):not(.bar)`_) 
I believe we can leave it for now, maybe changing it a bit to point out that for the browsers that still don't support it one could use the following syntax.
Would be happy to know other opinions.