Improving accuracy of `:nth-child()` example text
The current example text makes it sound like, in a situation where there were multiple sibling elements of different types, the second *list* element would be selected; when actually, if there were multiple types of elements at the same level and the second one was *not* a `<li>`, then nothing would be selected. I attempted to clarify the comment in the example, and added a note on the point below.