Remove "best practice" paragraph
Whether it makes more sense to separate `<link>` elements from the body content depends on the actual scenario. In some cases, it makes sense and might be preferable to use `<link>` elements in the `<body>`. We should avoid making judgment calls about whether something is best practice or not.

However, this isn't a good practice to follow; it makes more sense to separate your `<link>` elements from your body content, putting them in the `<head>`.