Clarify "Converting mistakes into errors" example
I [think] I clarified the example of how strict mode handles 'accidental' global variable initialization. The previous example mentioned a 'misspelled variable,' but there was none present in the example, which could lead to confusion. This PR merely _adds_ a variable declaration that can be misspelled in the rest of the example. Finally, I updated the comments in the example to use an actual misspelling of the declared variable. Thanks for your awesome resources!