Clarify "Converting mistakes into errors" example (#15873)
* Clarify "Converting mistakes into errors" example

I [think] I clarified the example of how strict mode handles 'accidental' global variable initialization. The previous example mentioned a 'misspelled variable,' but there was none present in the example, which could lead to confusion. This PR merely _adds_ a variable declaration that can be misspelled in the rest of the example. Finally, I updated the comments in the example to use an actual misspelling of the declared variable. Thanks for your awesome resources!

* Change `var` to `let` per review comment

* Update files/en-us/web/javascript/reference/strict_mode/index.md

Co-authored-by: Onkar Ruikar <87750369+OnkarRuikar@users.noreply.github.com>

* Add clarification to example by specifying misspelling

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

Co-authored-by: Onkar Ruikar <87750369+OnkarRuikar@users.noreply.github.com>
Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>