Correct code indentation to match program flow
An `if` statement with no opening and closing brackets (`{ }`) confuses people that are not familiar with its semantics.  Indenting the code following the first command after the `if` statement makes it even more confusing, since the indentation's implication is contrary to the actual program flow.  Not to mention all other `if` statements have opening and closing brackets.