The problem with the quest.
Sometimes was create in the JS the variable for the section that is need always and not sometimes, always is propose an text for the section this isn't need because was write nothing in the html explicite in the JS that is empty isn't necessary what is need is make the variable and not write nothing in JS, also the problem to find the 4 numbers is that 48 is a very factorable number so I gonna have infinite combinations of 4 numbers. So I will not have a number for each but an array, in math if I have 4 incognitos I need have 4 mathematical sentences to have sure numbers and not an array for numbers for each, or would be like statistic, for statistics I could find all the numbers that will match with your request is that I didn't learn statistic.