Remove misleading statement about setters
The remark made it seem like `Object.assign({}, objectWithSetters)` copies property descriptors. That is not true and thus the real reason setters are not involved is because the object is never mutated. The next example after this that references Object.assign is also kind of strange but decided not to touch it.