Drop Specifications table from ALPN glossary entry
The ALPN glossary entry is the only entry in the Glossary with a Specifications
section/table. It’s really not necessary for a glossary entry to have a spec
table, so this change replaces the table with a simple inline hyperlink to
the spec. The change also makes a couple other minor fixes.